+++ tomcat/src/main/java/org/apache/coyote/view/ContentTypeConverter.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.view;
 
 import java.util.Map;
